requirements.txt
langchain~=0.3.27
langgraph~=0.6.4
langchain-openai~=0.3.29
mem0ai
typing_extensions
python-dotenv~=1.1.1
typing_extensions~=4.14.1
langchain-core~=0.3.74
autogen~=0.9.9
mcp~=1.13.0
bs4
langchain-mcp-adapters
grandalf
graphviz
pygraphviz
ipython
requests~=2.32.5
typing_extensions~=4.14.1
openai~=1.100.2
typing_extensions~=4.14.1